create user show pages, different for teacher vs user
User show pages are used to show a students analytics and currently enrolled classes.
A student can see all the classes they are enrolled in and their submissions for those classes.
A teacher can only see the students progress in the classes that they administer.
This commit also updates the db/seed.rb file to have more interactions to reflect the above pages.