Text Editor Integration Complete
Updated views to show editor if student is logged in and on the assignment view page. If the assignment has already been submitted the editor is populated with the existing content and updates it upon submission. Rewrote submission controller methods create and update to accept values from editor. Editor is loaded from partial.